pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-alpaca-test-trade"
version = "0.1.0"
description = "Portable paper-trading-first Alpaca research, backtesting, and execution lab."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alpaca-py>=0.33.0",
  "matplotlib>=3.9",
  "numpy>=2.1",
  "pandas>=2.2",
  "pyarrow>=17.0",
  "pydantic>=2.8",
  "pyyaml>=6.0.2",
  "python-dotenv>=1.0.1",
  "requests>=2.32",
  "rich>=13.9.0",
  "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.10.0",
  "mypy>=1.11",
  "pytest>=8.3",
  "ruff>=0.6.9",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["alpaca_lab*"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["B008"]

[tool.ruff.lint.per-file-ignores]
"scripts/*.py" = ["E402"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
